fix: 3D scan vispy IndexError + detector pixel size float step
ImageWidget.setImage: set layer.scale before layer.data so napari's
internal _world_to_layer_units_scale is already the correct length when
dims_displayed is indexed during the data-change refresh.  The previous
order caused IndexError on the first 3D frame (old 2D scale still on
the layer when napari fired the display update).

SettingsWidget DetectorNumberParameter: add step=0.001 so the spinbox
arrows/scroll wheel move in 1nm (µm unit) or 1ms (s unit) increments
instead of defaulting to step=1.0 which effectively forced integer-only
editing via the controls.

DetectorManager: correct type annotations List[int] → List[float] for
scale and pixelSizeUm (they always contained floats; the wrong hint was
misleading and could encourage int coercion in calling code).

+++ imswitch/imcontrol/model/managers/detectors/DetectorManager.py
@@ -207,7 +207,7 @@ class DetectorManager(SignalInterface):
         return self.__forFocusLock
 
     @property
-    def scale(self) -> List[int]:
+    def scale(self) -> List[float]:
         """ The pixel sizes in micrometers, all axes, in the format high dim
         to low dim (ex. [..., 'Z', 'Y', 'X']). Override in managers handling
         >3 dim images (e.g. APDManager). """
@@ -215,7 +215,7 @@ class DetectorManager(SignalInterface):
 
     @property
     @abstractmethod
-    def pixelSizeUm(self) -> List[int]:
+    def pixelSizeUm(self) -> List[float]:
         """ The pixel size in micrometers, in 3D, in the format
         ``[Z, Y, X]``. Non-scanned ``Z`` set to 1. """
         pass
